Rename group functions and related refactoring changes
